+++ api-svc/spring/src/main/java/raido/apisvc/service/auth/RaidV2AppUserApiTokenService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Component;
 import raido.apisvc.repository.AppUserRepository;
 import raido.apisvc.spring.config.environment.RaidV2AppUserAuthProps;
 import raido.apisvc.spring.security.raidv2.ApiToken;
+import raido.apisvc.spring.security.raidv2.UnapprovedUserApiToken;
 import raido.apisvc.util.Guard;
 import raido.apisvc.util.Log;
 
@@ -18,7 +19,7 @@ import java.util.Optional;
 
 import static java.util.Optional.of;
 import static org.eclipse.jetty.util.TypeUtil.isFalse;
-import static raido.apisvc.service.auth.UnapprovedUserApiToken.UnapprovedUserApiTokenBuilder.anUnapprovedUserApiToken;
+import static raido.apisvc.spring.security.raidv2.UnapprovedUserApiToken.UnapprovedUserApiTokenBuilder.anUnapprovedUserApiToken;
 import static raido.apisvc.spring.security.raidv2.ApiToken.ApiTokenBuilder.anApiToken;
 import static raido.apisvc.util.ExceptionUtil.authFailed;
 import static raido.apisvc.util.ExceptionUtil.wrapException;